Fix logic add map for address
