+++ app/controllers/clients_controller.rb
@@ -51,16 +51,6 @@ class ClientsController < ApplicationController
     end
   end
 
-  def show_address
-    @address = Geocoder.search(params[:position])
-    render json: { html: @address }
-  end
-
-  def handler_address
-    @address = Geocoder.search(params[:address])
-    render json: { html: @address }
-  end
-
   private
 
   def set_client
