+++ app/controllers/employees_controller.rb
@@ -45,16 +45,6 @@ class EmployeesController < ApplicationController
     end
   end
 
-  def show_address
-    @address = Geocoder.search(params[:position])
-    render json: { html: @address }
-  end
-
-  def handler_address
-    @address = Geocoder.search(params[:address])
-    render json: { html: @address }
-  end
-
   private
 
   def set_employee
