+++ app/javascript/packs/application.js
@@ -18,6 +18,5 @@ const images = require.context('../images', true)
 const imagePath = (name) => images(name, true)
 import '../application';
 
-
-require("./employee")
+require("./geocoder")
 require("./item")
